+++ app/src/main/java/pk/sufiishq/app/core/event/events/GlobalEvents.kt
@@ -8,6 +8,7 @@ abstract class GlobalEvents : Event(GlobalEventHandler::class) {
 
     class ShareKalam(val kalam: Kalam, val context: Context) : GlobalEvents()
     class ShareApp(val context: Context) : GlobalEvents()
-    class ShowUpdateButton(val isShow: Boolean): GlobalEvents()
-    object StartUpdateFlow: GlobalEvents()
+    class ShowUpdateButton(val isShow: Boolean) : GlobalEvents()
+    class OpenFacebookGroup(val context: Context, val groupUrl: String) : GlobalEvents()
+    object StartUpdateFlow : GlobalEvents()
 }